requirements.txt
Package                   Version
------------------------- ---------
asgiref                   3.5.0
Authlib                   0.15.5
black                     21.12b0
certifi                   2021.10.8
cffi                      1.15.0
charset-normalizer        2.0.10
click                     8.0.3
coreapi                   2.3.3
coreschema                0.0.4
cryptography              36.0.1
Django                    4.0.1
django-cors-headers       3.11.0
django-dotenv             1.4.2
django-stubs              1.9.0
django-stubs-ext          0.3.1
djangorestframework       3.13.1
djangorestframework-stubs 1.4.0
idna                      3.3
itypes                    1.2.0
Jinja2                    3.0.3
MarkupSafe                2.0.1
mypy                      0.931
mypy-extensions           0.4.3
pathspec                  0.9.0
pip                       21.3.1
platformdirs              2.4.1
pycparser                 2.21
PyJWT                     2.3.0
pytz                      2021.3
requests                  2.27.1
setuptools                58.1.0
sqlparse                  0.4.2
toml                      0.10.2
tomli                     1.2.3
types-pytz                2021.3.4
types-PyYAML              6.0.3
types-requests            2.27.7
types-urllib3             1.26.7
typing_extensions         4.0.1
uritemplate               4.1.1
urllib3                   1.26.8
